requirements.txt
absl-py==2.1.0
ale-py==0.8.1
asttokens==2.4.1
attrs==24.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.2.0
grpcio==1.63.0
gym==0.25.2
gym-notices==0.0.8
gym-super-mario-bros==7.4.0
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mistune==3.0.2
mkl==2021.4.0
moviepy==1.0.3
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nes-py==8.2.1
networkx==3.2.1
numpy==1.23.5
opencv-python==4.9.0.80
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==5.26.1
pure_eval==0.2.3
pygame==2.1.0
pyglet==1.5.21
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.2
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tinycss2==1.4.0
torch==2.3.0+cu121
torchaudio==2.3.0+cu121
torchvision==0.18.0+cu121
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
yarg==0.1.9
